Changes to be committed:

+++ database/migrations/2019_10_22_141706_create_employees_table.php
@@ -17,7 +17,7 @@ class CreateEmployeesTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('job_title');
-            $table->decimal('salary');
+            $table->double('salary',15,2);
             $table->string('email')->unique();
             $table->string('password');
             $table->rememberToken();
